Separate data from story
