+++ src/stories/cartesian/brush/Brush.stories.tsx
@@ -1,0 +1,17 @@
+import React from 'react';
+import BrushComponent from './BrushComponent';
+import { dateData } from '../../data';
+
+export default {
+  title: 'Recharts/Brush',
+  component: BrushComponent,
+};
+
+function Template(args: any) {
+  return <BrushComponent {...args} />;
+}
+
+export const Simple = Template.bind({});
+Simple.args = {
+  data: dateData,
+};
